+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.0.2
+
+- moved ~/.emu-tools/.env to ~/.config/emu-tools/.env
+- reduced duplicated code in install.sh
+
 ## 1.0.1
 
 - fix path validation in media_cleaner.sh, to avoid removing files if the path are incorrect.